Add files via upload
Cleaned up the querying function. There is a dictionary called "allParams" which is currently in the query function. This branch will contain only a modified query function, however there is another branch where I have cleaned up the code to make it avoid repetition (e.g. making "allParams" a global variable and using it in the other functions as well) and offer utility.